services/backend_for_frontend/requirements.txt
# Python

# Django
psycopg2==2.9.9
Django==5.0.6
# Django Rest Framework
djangorestframework==3.15.0
djangorestframework-api-key==3.0.0